model/inference.py
import numpy as np
from PIL import Image
# import tensorflow as tf

def preprocess_image(image_path):
    """
    Prepares an image for the ML model.
    """
    img = Image.open(image_path).resize((224, 224))
    img_array = np.array(img) / 255.0
    return np.expand_dims(img_array, axis=0)

def predict_breed(model, image_path):
    """
    Conceptual inference function.
    """
    # processed_img = preprocess_image(image_path)
    # predictions = model.predict(processed_img)
    # return np.argmax(predictions)
    return "Gir Cow" # Placeholder

if __name__ == "__main__":
    print("Conceptual inference script loaded.")
